set Employee role condition
